# Conflicts:
#	src/main/java/com/example/touristguideapigruppe/repositories/TouristRepository.java

+++ src/main/java/com/example/touristguideapigruppe/controllers/TouristController.java
@@ -1,9 +1,10 @@
 package com.example.touristguideapigruppe.controllers;
 
+import com.example.touristguideapigruppe.exceptions.NotFoundException;
+import com.example.touristguideapigruppe.exceptions.NullException;
+import com.example.touristguideapigruppe.exceptions.UnkownErrorException;
 import com.example.touristguideapigruppe.models.TouristAttraction;
 import com.example.touristguideapigruppe.services.TouristService;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -13,12 +14,13 @@ import java.util.List;
 @Controller
 @RequestMapping("/attractions")
 public class TouristController {
-    private TouristService touristService;
+    private final TouristService touristService;
 
     public TouristController(TouristService touristService) {
         this.touristService = touristService;
     }
 
+    // hent "index" til attraction
     @GetMapping("")
     public String getAttractions(Model model) {
         List<TouristAttraction> attractions = touristService.getAttractions();
@@ -26,42 +28,65 @@ public class TouristController {
         return "attractions";
     }
 
+    // hente specifik
     @GetMapping("{name}")
-    public String getByName(@PathVariable String name, Model model) {
+    public String getByName(@PathVariable String name, Model model) throws NotFoundException {
         model.addAttribute("specific", touristService.getByName(name));
         return "attraction-information";
     }
 
+    // henter add layout
     @GetMapping("insertAttraction")
-    public String getHandleAttraction(Model model){
+    public String getAddAttraction(Model model) {
         model.addAttribute("attraction",
                 new TouristAttraction());
         return "add-attraction";
     }
 
+    // tilføj metoden
     @PostMapping("addAttraction")
-    public String addAttraction(@ModelAttribute("attraction")TouristAttraction attraction) {
-        touristService.addAttraction(attraction);
+    public String addAttraction(@ModelAttribute("attraction") TouristAttraction attraction) {
+        if (attraction == null) throw new NullException();
+
+        TouristAttraction result = touristService.addAttraction(attraction);
+        if (result == null) throw new UnkownErrorException();
+
+        // laver en 302 response sådan, at ikke kan poste det samme igen.
         return "redirect:/attractions";
     }
 
+    // henter update/delete layout
     @GetMapping("handleAttraction/{name}")
-    public String getHandleAttraction(@PathVariable("name") String name, Model model){
+    public String getHandleAttraction(@PathVariable("name") String name, Model model) throws NotFoundException {
+        if (name == null) throw new NullException();
+
         TouristAttraction result = touristService.getByName(name);
         model.addAttribute("attraction",
                 result);
         return "handle-attraction";
     }
 
+    // opdater metoden
     @PostMapping("updateAttraction")
-    public String updateAttraction(@ModelAttribute("attraction")TouristAttraction newAttraction) {
-        touristService.updateAttraction(newAttraction);
+    public String updateAttraction(@ModelAttribute("attraction") TouristAttraction newAttraction) throws NotFoundException {
+        if (newAttraction == null) throw new NullException();
+
+        TouristAttraction result = touristService.updateAttraction(newAttraction);
+        if (result == null) throw new UnkownErrorException();
+
+        // laver en 302 response sådan, at ikke kan poste det samme igen.
         return "redirect:/attractions";
     }
 
+    // delete metoden
     @PostMapping("delete/{name}")
-    public String deleteAttraction(@PathVariable String name) {
+    public String deleteAttraction(@PathVariable String name) throws NotFoundException {
+        if (name == null) throw new NullException();
+
         TouristAttraction result = touristService.deleteAttraction(name);
+        if (result == null) throw new UnkownErrorException();
+
+        // laver en 302 response sådan, at ikke kan poste det samme igen.
         return "redirect:/attractions";
     }
 }
